fix(nav): client boundary & remove server component onClick
Fix runtime error: "Event handlers cannot be passed to Client Component props"

Changes:
- Extract SkipLink to separate client component (accessibility)
- Add usePathname() + useEffect to auto-close mobile menu on navigation
- Remove problematic onClick handlers from server-side layout.tsx
- Add missing usePageAnalytics hook

Runtime test: ✅ /cart loads 200 (no more 500 errors)
Build test: ✅ Compilation successful (1309ms)

Fixes blocking issue for PR merge sequence.
